+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.1.0 - TBD
+
+### Features
+
+- add `generate` and `generator` aliases for `enumerate` and `enumerator`
+
+### Fixes
+
+- fix type inference for alphabet elements
+
+### Changes
+
+- add the invalid value to type errors
+- documentation tweaks
+
 ## 0.0.1 - 2020-12-29
 
 - initial release